Test reading plaintext footer files with encryption both disabled and enabled
