+++ parquet/tests/arrow_reader/encryption.rs
@@ -18,13 +18,10 @@
 //! This module contains tests for reading encrypted Parquet files with the Arrow API
 
 use crate::encryption_util::verify_encryption_test_data;
-use arrow_array::cast::AsArray;
-use arrow_array::{types, RecordBatch};
-use arrow_schema::ArrowError;
+use arrow_array::RecordBatch;
 use parquet::arrow::arrow_reader::{
     ArrowReaderMetadata, ArrowReaderOptions, ParquetRecordBatchReaderBuilder,
 };
-use parquet::arrow::ProjectionMask;
 use parquet::encryption::decrypt::FileDecryptionProperties;
 use std::fs::File;
 
@@ -107,60 +104,9 @@ fn test_non_uniform_encryption_disabled_aad_storage() {
 }
 
 #[test]
-fn test_non_uniform_encryption_plaintext_footer_without_decryption() {
-    let test_data = arrow::util::test_util::parquet_test_data();
-    let path = format!("{test_data}/encrypt_columns_plaintext_footer.parquet.encrypted");
-    let file = File::open(&path).unwrap();
-
-    let metadata = ArrowReaderMetadata::load(&file, Default::default()).unwrap();
-    let file_metadata = metadata.metadata().file_metadata();
-
-    assert_eq!(file_metadata.num_rows(), 50);
-    assert_eq!(file_metadata.schema_descr().num_columns(), 8);
-    assert_eq!(
-        file_metadata.created_by().unwrap(),
-        "parquet-cpp-arrow version 19.0.0-SNAPSHOT"
-    );
-
-    metadata.metadata().row_groups().iter().for_each(|rg| {
-        assert_eq!(rg.num_columns(), 8);
-        assert_eq!(rg.num_rows(), 50);
-    });
-
-    // Should be able to read unencrypted columns. Test reading one column.
-    let builder = ParquetRecordBatchReaderBuilder::try_new(file).unwrap();
-    let mask = ProjectionMask::leaves(builder.parquet_schema(), [1]);
-    let record_reader = builder.with_projection(mask).build().unwrap();
-
-    let mut row_count = 0;
-    for batch in record_reader {
-        let batch = batch.unwrap();
-        row_count += batch.num_rows();
-
-        let time_col = batch
-            .column(0)
-            .as_primitive::<types::Time32MillisecondType>();
-        for (i, x) in time_col.iter().enumerate() {
-            assert_eq!(x.unwrap(), i as i32);
-        }
-    }
-
-    assert_eq!(row_count, file_metadata.num_rows() as usize);
-
-    // Reading an encrypted column should fail
-    let file = File::open(&path).unwrap();
-    let builder = ParquetRecordBatchReaderBuilder::try_new(file).unwrap();
-    let mask = ProjectionMask::leaves(builder.parquet_schema(), [4]);
-    let mut record_reader = builder.with_projection(mask).build().unwrap();
-
-    match record_reader.next() {
-        Some(Err(ArrowError::ParquetError(s))) => {
-            assert!(s.contains("protocol error"));
-        }
-        _ => {
-            panic!("Expected ArrowError::ParquetError");
-        }
-    };
+#[cfg(feature = "snap")]
+fn test_plaintext_footer_read_without_decryption() {
+    crate::encryption_common::read_plaintext_footer_file_without_decryption_properties();
 }
 
 #[test]
